requirements.txt
asn1crypto==0.24.0
boto3==1.9.185
botocore==1.12.185
cffi==1.12.3
cryptography==2.7
dj-database-url==0.5.0
Django==2.2.3
django-heroku==0.3.1
django-rest-knox==4.1.0
django-storages==1.7.1
djangorestframework==3.9.4
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.4
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.1
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.2
psycopg2-binary==2.8.3
